# docs/openapi.yaml
...
openapi: 3.0.3
info:
  title: TravelCash API
  version: 2.0.0
  description: >-
    API REST do **TravelCash — Controle Financeiro para Viagens** (FIAP • CP2).


    **Padrão de resposta**: `{ success, data, meta? }` em sucesso e `{
    success:false, error:{ code, message, details? } }` em erro.


    **Listagens** aceitam `page`, `limit` (máx. 100, padrão 20), `sort` (prefixo
    `-` = decrescente) e `search`; filtros específicos de cada recurso estão
    descritos nos parâmetros.


    **IA (LLM)**: endpoints em *IA* usam Gemini ou Claude quando há chave
    configurada; caso contrário respondem com regras locais (`source:
    "regras"`).
  license:
    name: Uso acadêmico — FIAP
servers:
  - url: /
    description: Servidor atual
...
paths:
...
  /api/dashboard:
    get:
      tags:
        - Dashboard
      summary: Dashboard consolidado de uma viagem
      description: Indicadores, gastos por categoria, evolução diária, projeção e
        alertas automáticos — agregados em SQL.
      parameters:
        - name: tripId
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      trip:
                        $ref: "#/components/schemas/Trips"
                      budget:
                        $ref: "#/components/schemas/Budgets"
                      summary:
                        type: object
                        properties:
                          totalBudget:
                            type: number
                          totalSpent:
                            type: number
                          remaining:
                            type: number
                          utilization:
                            type: number
                          dailyAverage:
                            type: number
                          days:
                            type: number
                          expenseCount:
                            type: number
                          largestExpense:
                            type: number
                          projectedTotal:
                            type: number
                      categories:
                        type: array
                        items:
                          type: object
                          properties:
                            category:
                              type: string
                            color:
                              type: string
                            total:
                              type: number
                            share:
                              type: number
                            count:
                              type: integer
                      categoryTotals:
                        type: object
                        additionalProperties:
                          type: number
                        description: Formato legado (CP1)
                      dailySpending:
                        type: array
                        items:
                          type: object
                          properties:
                            date:
                              type: string
                              format: date
                            total:
                              type: number
                            count:
                              type: integer
                      alerts:
                        type: array
                        description: Alertas automáticos (auto=true, calculados por regras) + manuais
                        items:
                          type: object
                      expenses:
                        type: array
                        items:
                          $ref: "#/components/schemas/Expenses"
                      itinerary:
                        type: array
                        items:
                          $ref: "#/components/schemas/Itinerary"
                      goals:
                        type: array
                        items:
                          $ref: "#/components/schemas/Goals"
        "404":
          $ref: "#/components/responses/NotFound"
      operationId: getDashboard
      security: []
  /api/trips/{id}/dashboard:
    get:
      tags:
        - Dashboard
      summary: Dashboard da viagem (rota REST)
      parameters:
        - $ref: "#/components/parameters/id"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      trip:
                        $ref: "#/components/schemas/Trips"
                      budget:
                        $ref: "#/components/schemas/Budgets"
                      summary:
                        type: object
                        properties:
                          totalBudget:
                            type: number
                          totalSpent:
                            type: number
                          remaining:
                            type: number
                          utilization:
                            type: number
                          dailyAverage:
                            type: number
                          days:
                            type: number
                          expenseCount:
                            type: number
                          largestExpense:
                            type: number
                          projectedTotal:
                            type: number
                      categories:
                        type: array
                        items:
                          type: object
                          properties:
                            category:
                              type: string
                            color:
                              type: string
                            total:
                              type: number
                            share:
                              type: number
                            count:
                              type: integer
                      categoryTotals:
                        type: object
                        additionalProperties:
                          type: number
                        description: Formato legado (CP1)
                      dailySpending:
                        type: array
                        items:
                          type: object
                          properties:
                            date:
                              type: string
                              format: date
                            total:
                              type: number
                            count:
                              type: integer
                      alerts:
                        type: array
                        description: Alertas automáticos (auto=true, calculados por regras) + manuais
                        items:
                          type: object
                      expenses:
                        type: array
                        items:
                          $ref: "#/components/schemas/Expenses"
                      itinerary:
                        type: array
                        items:
                          $ref: "#/components/schemas/Itinerary"
                      goals:
                        type: array
                        items:
                          $ref: "#/components/schemas/Goals"
        "404":
          $ref: "#/components/responses/NotFound"
      operationId: getTripsByIdDashboard
      security: []
...
  /api/ai/classify-expense:
    post:
      tags:
        - IA
      summary: Sugerir categoria de uma despesa (LLM)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - description
              properties:
                description:
                  type: string
                  maxLength: 200
                amount:
                  type: number
            example:
              description: Uber até o aeroporto
              amount: 85
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/Classification"
        "400":
          $ref: "#/components/responses/BadRequest"
        "429":
          $ref: "#/components/responses/RateLimited"
      operationId: postAiClassifyExpense
      security: []
  /api/trips/{id}/insights:
    get:
      tags:
        - IA
      summary: Análise financeira da viagem gerada por LLM
      description: Envia ao modelo apenas números agregados (sem nome/e-mail).
        Respostas ficam em cache por 10 min; use `refresh=true` para forçar nova
        geração.
      parameters:
        - $ref: "#/components/parameters/id"
        - name: refresh
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/Insights"
        "404":
          $ref: "#/components/responses/NotFound"
        "429":
          $ref: "#/components/responses/RateLimited"
      operationId: getTripsByIdInsights
      security: []
  /api/ai/insights:
    get:
      tags:
        - IA
      summary: Análise financeira (variação por query string)
      parameters:
        - name: tripId
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
        - name: refresh
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/Insights"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
      operationId: getAiInsights
      security: []
components:
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
...
  schemas:
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Dados inválidos na requisição.
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    Insights:
      type: object
      properties:
        tripId:
          type: integer
        source:
          type: string
          enum:
            - llm
            - regras
        model:
          type: string
          nullable: true
          example: gemini-2.5-flash
        cached:
          type: boolean
        generatedAt:
          type: string
          format: date-time
        fallbackReason:
          type: string
        dataSent:
          type: object
          description: Dados agregados enviados ao modelo (sem dados pessoais)
        insights:
          type: object
          properties:
            resumo:
              type: string
            saude:
              type: string
              enum:
                - boa
                - atencao
                - critica
            pontosDeAtencao:
              type: array
              items:
                type: string
            recomendacoes:
              type: array
              items:
                type: string
            previsao:
              type: string
    Classification:
      type: object
      properties:
        category:
          type: string
          example: Alimentação
        confidence:
          type: number
          example: 0.92
        reason:
          type: string
          example: Refeição em restaurante.
        source:
          type: string
          enum:
            - llm
            - regras
        model:
          type: string
          nullable: true
...
    Trips:
      type: object
      required:
        - name
        - startDate
        - endDate
      properties:
        id:
          type: integer
          readOnly: true
        userId:
          type: integer
          minimum: 1
          description: ID de usuário (FK)
        name:
          type: string
          maxLength: 120
        destination:
          type: string
          maxLength: 120
        startDate:
          type: string
          format: date
          example: 2025-06-01
        endDate:
          type: string
          format: date
          example: 2025-06-01
        travelers:
          type: integer
          minimum: 1
          maximum: 50
        status:
          type: string
          enum:
            - Planejamento
            - Em andamento
            - Concluída
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      example:
        id: 1
        name: Orlando - Férias 2025
        destination: Orlando, EUA
        startDate: 2025-06-01
        endDate: 2025-06-15
        travelers: 2
        status: Em andamento
...
    Budgets:
      type: object
      required:
        - tripId
        - totalBudget
      properties:
        id:
          type: integer
          readOnly: true
        tripId:
          type: integer
          minimum: 1
          description: ID de viagem (FK)
        totalBudget:
          type: number
          minimum: 0
        dailyBudget:
          type: number
          minimum: 0
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      example:
        id: 1
        tripId: 2
        totalBudget: 15000
        dailyBudget: 300
...
    Expenses:
      type: object
      required:
        - tripId
        - category
        - description
        - amount
        - date
      properties:
        id:
          type: integer
          readOnly: true
        tripId:
          type: integer
          minimum: 1
          description: ID de viagem (FK)
        category:
          type: string
          description: Nome da categoria (GET /api/categories). Em POST, omitir ou enviar
            "auto" faz a IA classificar.
        description:
          type: string
          maxLength: 200
        amount:
          type: number
          minimum: 0.01
          maximum: 1000000
        date:
          type: string
          format: date
          example: 2025-06-01
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        categoryId:
          type: integer
          readOnly: true
        categoryColor:
          type: string
          readOnly: true
      example:
        id: 1
        tripId: 1
        category: Alimentação
        description: Jantar no restaurante
        amount: 120.5
        date: 2025-06-03
...
    Itinerary:
      type: object
      required:
        - tripId
        - date
        - title
      properties:
        id:
          type: integer
          readOnly: true
        tripId:
          type: integer
          minimum: 1
          description: ID de viagem (FK)
        date:
          type: string
          format: date
          example: 2025-06-01
        title:
          type: string
          maxLength: 160
        location:
          type: string
          maxLength: 160
        estimatedCost:
          type: number
          minimum: 0
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
      example:
        id: 1
        tripId: 1
        date: 2025-06-04
        title: Magic Kingdom
        location: Walt Disney World
        estimatedCost: 680
...
    Goals:
      type: object
      required:
        - name
        - targetAmount
      properties:
        id:
          type: integer
          readOnly: true
        userId:
          type: integer
          minimum: 1
          description: ID de usuário (FK)
        name:
          type: string
          maxLength: 120
        targetAmount:
          type: number
          minimum: 0.01
        currentAmount:
          type: number
          minimum: 0
        createdAt:
          type: string
          format: date-time
          readOnly: true
        updatedAt:
          type: string
          format: date-time
          readOnly: true
        progress:
          type: number
          readOnly: true
          description: Calculado (currentAmount / targetAmount)
      example:
        id: 1
        name: Viagem Europa 2026
        targetAmount: 15000
        currentAmount: 6250
...
  responses:
    BadRequest:
      description: Requisição inválida (validação/JSON/parâmetros)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Não encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimited:
      description: Limite de requisições excedido
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"